Compilador/proyect/requirements.txt
Django 
gunicorn
djangorestframework
django-cors-headers
antlr4-tools
antlr4-python3-runtime==4.13.2